Clean up URI handling by giving the Inliner a default base URI
There are now 3 levels of URIs that resolve when a href attribute is
dereferenced:

- The href's URI-reference
- The href's element's base URI-reference
- The Inliner's default base URI

The Inliner's default base URI defaults to file:<current dir>, which
has the effect of transparently resolving relative/paths/like/this to
the file: URIs which the FilesystemUrlHandler handles.

It also means that URIs passed to can_handle() methods are always URIs
(rather than URI-references, which could be scheme-less).