fix: handle indirectly-included components when overriding
Previously when inlining <include> elements, the include's override
components were only able to override components that directly occurred
in the included schema. If the included schema itself didn't directly
contain an overridden <define> or <start> element, but instead included
another schema that did contain an overridden element, rnginline would
fail to find the indirectly-included element(s) and would fail with an
error. (Because overrides must match elements in the included schema.)

This occurred because of the deferred XML inclusion workaround we use to
lazily merge XML trees. Rather than inserting XML elements from
referenced schemas directly into the referencing schema's XML tree, we
instead maintain a tree of the positions that included elements need to
placed at when the final merging is performed.

When we searched for overridden components, we only looked in the
immediate XML root of the schema that was included, without descending
into any pending, indirect XML trees. To fix this, we now fully traverse
the deferred XML inclusion tree when searching for overridden components.

See issue #5.